requirements.txt
pandas>=2.2.0
numpy>=1.26.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pyyaml>=6.0.1
requests>=2.31.0
tqdm>=4.66.0
discord.py>=2.4.0
